# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A read-only Aave V3 data monitoring service",
        "title": "aavepulse",
        "contact": {},
        "version": "v0.1.0"
    },
    "paths": {
...
        "/api/v1/menus": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Query menu tree data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code path of menu (like xxx.xxx.xxx)",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Name of menu",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Whether to include menu resources",
                        "name": "includeResources",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.Menu"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Create menu record",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.MenuForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.Menu"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/menus/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Get menu record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.Menu"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Update menu record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.MenuForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "MenuAPI"
                ],
                "summary": "Delete menu record by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "unique id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errors.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "schema.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "Child menus",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Menu"
                    }
                },
                "code": {
                    "description": "Code of menu (unique for each level)",
                    "type": "string"
                },
                "created_at": {
                    "description": "Create time",
                    "type": "string"
                },
                "description": {
                    "description": "Details about menu",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "name": {
                    "description": "Display name of menu",
                    "type": "string"
                },
                "parent_id": {
                    "description": "Parent ID (From Menu.ID)",
                    "type": "string"
                },
                "parent_path": {
                    "description": "Parent path (split by .)",
                    "type": "string"
                },
                "path": {
                    "description": "Access path of menu",
                    "type": "string"
                },
                "properties": {
                    "description": "Properties of menu (JSON)",
                    "type": "string"
                },
                "resources": {
                    "description": "Resources of menu",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuResource"
                    }
                },
                "sequence": {
                    "description": "Sequence for sorting (Order by desc)",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of menu (enabled, disabled)",
                    "type": "string"
                },
                "type": {
                    "description": "Type of menu (page, button)",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
        "schema.MenuForm": {
            "type": "object",
            "required": [
                "code",
                "name",
                "status",
                "type"
            ],
            "properties": {
                "code": {
                    "description": "Code of menu (unique for each level)",
                    "type": "string",
                    "maxLength": 32
                },
                "description": {
                    "description": "Details about menu",
                    "type": "string"
                },
                "name": {
                    "description": "Display name of menu",
                    "type": "string",
                    "maxLength": 128
                },
                "parent_id": {
                    "description": "Parent ID (From Menu.ID)",
                    "type": "string"
                },
                "path": {
                    "description": "Access path of menu",
                    "type": "string"
                },
                "properties": {
                    "description": "Properties of menu (JSON)",
                    "type": "string"
                },
                "resources": {
                    "description": "Resources of menu",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.MenuResource"
                    }
                },
                "sequence": {
                    "description": "Sequence for sorting (Order by desc)",
                    "type": "integer"
                },
                "status": {
                    "description": "Status of menu (enabled, disabled)",
                    "type": "string",
                    "enum": [
                        "disabled",
                        "enabled"
                    ]
                },
                "type": {
                    "description": "Type of menu (page, button)",
                    "type": "string",
                    "enum": [
                        "page",
                        "button"
                    ]
                }
            }
        },
        "schema.MenuResource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Create time",
                    "type": "string"
                },
                "id": {
                    "description": "Unique ID",
                    "type": "string"
                },
                "menu_id": {
                    "description": "From Menu.ID",
                    "type": "string"
                },
                "method": {
                    "description": "HTTP method",
                    "type": "string"
                },
                "path": {
                    "description": "API request path (e.g. /api/v1/users/:id)",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Update time",
                    "type": "string"
                }
            }
        },
...
        "util.ResponseResult": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/errors.Error"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}